requirements.txt
groq==0.11.0 
cohere==5.11.2 
python-dotenv==1.0.1 
requests==2.32.3
fastapi==0.115.4 